Disable libuv signal handling
The signal handlers set by libuv were not be unset when calling
`uv_loop_close()`, so would cause crashes when fs was unloaded and the R
process was later forked by processx, parallel, sys or other packages.

Disabling the signal initialization fixes the issue, as we do not use
any signal handlers in fs.