Return uncolored for malformed LS_COLORS
We could try to patch up the empty case as shown in the issue, but I
think it is better to just bail and give up coloring if something seems
wrong. I decided not to issue a warning for this case, if it comes up
again we can think about adding one.

Fixes #135