Test that path_rel() can be reversed by path_abs().
This is related to Issues #144 and #174.

On macOS, the temp directory is a symlink, which wouldn't be resolved
unless it existed. Now that path_real() resolves symlinks in the path
even if the target destination does not exist, this test should pass
even on macOS machines with symlinked temp directories.

On winbuilder, getwd() returns d:/ and tempdir() returns D:/, which
caused problems with path_rel(). Now that tidy paths always capitalize
the Windows drive, this test should pass on winbuilder.